requirements.txt
Django==1.4.2
South==0.7.6
django-braces==0.2.1
django-taggit==0.9.3
psycopg2==2.4.5
wsgiref==0.1.2
